model_consumer/requirements.txt
kafka-python==1.4.7
pymongo==3.10.1